% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shRNA_design.R
\name{OligoDesigner}
\alias{OligoDesigner}
\title{Generate shRNAs against circRNA back-splice junctions}
\usage{
OligoDesigner(
  sp = "dm6",
  input_coordinates = "circs_totest.txt",
  writetab = T,
  output = "shRNAs.tsv",
  shift = T,
  circRNAseq = T
)
}
\arguments{
\item{sp}{character with Name of the genome to use, available are: dm6, dm3, hg19, hg38, mm10, rn4. Default is dm6}

\item{writetab}{a boolean, if writetab=T, then the function will not write any table, instead it will return the table to R}

\item{output}{if writetab = T, it will create the output in this path, default is "shRNAs.tsv"}

\item{shift}{boolean, if shift=T, return also 5´and 3´shift sequences}

\item{circRNAseq}{boolean, if circRNAseq=T, the final table has the putative circRNA sequence asuming there is no internal alternative splicing (ie. all the possible internal exons one after the other)}

\item{input_coordinate}{document with the coordinate of the circRNAs to generate shRNAs in TSV format tab seppatated. It requires a column with Chromosome, Start and End.}
}
\value{
if writetab=F, It generates a DataFrame with the original table with the shRNA deisng appended
}
\description{
This script contains ONE funciton that will read a table with circRNA/exon-exon splicing junction coordinates and genenames in the format: Name Chr Start End.
Optional:a column named Strand
So far the following species are available: fly: dm3, dm6 and human: hs19, mice: mm10 and rat: rn4
}
\examples{
#There are examples in the folder names "../Test". To run it:
OligoDesigner(input_coordinates = "../test/circs_totest.txt", writetab=F)


}
